2. Logical restructurization of project
